app/order-service/requirements.txt
fastapi
uvicorn
httpx
pydantic
asyncio
python-jose[cryptography] 
prometheus-fastapi-instrumentator